Remove summary screen, move listing creation calls to Services step
